C printf extra funtions commit

+++ holberton.h
@@ -7,9 +7,9 @@
 #include <unistd.h>
 
 /**
-* struct conv_spec- Struct function that matches the @spec
+* struct conv_spec- Struct function that matches the spec
 * @spec: Conversion specifier symbol
-* @func: Associated function 
+* @func: Associated function
 */
 
 typedef struct conv_spec
@@ -39,7 +39,7 @@ int _printf_conv_spec_R(va_list args);
 
 int get_binary(unsigned int n);
 int get_octal(unsigned int n);
-int get_lower_hexa(unsigned int n);
+int get_lower_hexa(long n);
 int get_upper_hexa(unsigned int n);
 
 #endif
